requirements.txt
appdirs==1.4.4
audioread==2.1.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cycler==0.11.0
decorator==5.1.0
docopt==0.6.2
easydict==1.9
fonttools==4.28.3
future==0.18.2
googledrivedownloader==0.4
idna==3.3
isodate==0.6.0
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.2
librosa==0.8.0
llvmlite==0.37.0
lmdb==0.94
MarkupSafe==2.0.1
matplotlib==3.5.0
networkx==2.6.3
numba==0.54.1
numpy==1.20.3
opencv-python==4.5.4.60
packaging==21.3
pandas==1.3.4
Pillow==8.4.0
pipreqs==0.4.13
pooch==1.5.2
protobuf==3.19.1
pycparser==2.21
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
rdflib==6.0.2
requests==2.26.0
resampy==0.2.2
scikit-learn==1.0.1
scipy==1.7.3
setuptools-scm==6.3.2
six==1.16.0
SoundFile==0.10.3.post1
tensorboardX==2.1
threadpoolctl==3.0.0
tomli==1.2.2
torch==1.4.0
torch-cluster 
torch-geometric==2.0.2
torch-scatter 
torch-sparse
torch-spline-conv
torchaudio==0.4.0
torchvision==0.5.0
tqdm==4.62.3
typing_extensions==4.0.1
urllib3==1.26.7
yacs==0.1.8
yarg==0.1.9
